refactor: add folders, changed file locations
