+++ src/resources/auth/auth.module.ts
@@ -6,12 +6,12 @@ import { PassportModule } from '@nestjs/passport';
 
 import { JwtAuthGuard } from './guards-strategies/jwt-auth.guard';
 
-import { AuthService } from './auth.service';
+import { AuthService } from './services/auth.service';
 
 import { JwtStrategy } from './guards-strategies/jwt.strategy';
 import { UsersModule } from '../users/users.module';
 import { DrizzleModule } from 'src/infra/drizzle/drizzle.module';
-import { AuthController } from './auth.controller';
+import { AuthController } from './controllers/auth.controller';
 
 @Global()
 @Module({
