+++ src/resources/auth/guards-strategies/jwt-auth.guard.ts
@@ -1,7 +1,7 @@
 import { ExecutionContext, Injectable } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { AuthGuard } from '@nestjs/passport';
-import { IS_PUBLIC_KEY } from '../constants';
+import { IS_PUBLIC_KEY } from '../utils/constants';
 import { Strategies } from './types';
 
 @Injectable()
@@ -26,7 +26,6 @@ export class JwtAuthGuard extends AuthGuard(Strategies.JWT) {
   }
 
   getRequest(ctx: ExecutionContext) {
-    const request = ctx.switchToHttp().getRequest();
-    return request;
+    return ctx.switchToHttp().getRequest();
   }
 }
