+++ src/resources/auth/guards-strategies/jwt.strategy.ts
@@ -3,9 +3,9 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ConfigService } from '@nestjs/config';
 import { Payload, Strategies } from './types';
-import { UsersService } from 'src/resources/users/users.service';
+import { UsersService } from 'src/resources/users/services/users.service';
 import { UserResponse } from 'src/resources/users/dtos/responses';
-import { AuthService } from '../auth.service';
+import { AuthService } from '../services/auth.service';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, Strategies.JWT) {
