+++ src/resources/auth/models/auth-info.model.ts
@@ -1,0 +1,3 @@
+import { authInfo } from 'src/infra/drizzle/schema';
+
+export type AuthInfoModel = typeof authInfo.$inferInsert;
