+++ src/resources/users/models/user.model.ts
@@ -1,0 +1,3 @@
+import { users } from 'src/infra/drizzle/schema';
+
+export type UserModel = typeof users.$inferSelect;
