+++ src/resources/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module, forwardRef } from '@nestjs/common';
-import { UsersController } from './users.controller';
-import { UsersService } from './users.service';
-import { UserMapper } from './user.mapper';
+import { UsersController } from './controllers/users.controller';
+import { UsersService } from './services/users.service';
+import { UserMapper } from './services/user.mapper';
 import { DrizzleModule } from 'src/infra/drizzle/drizzle.module';
 import { AuthModule } from '../auth/auth.module';
